# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Teste Back-end PHP - Drugovich(V1)",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "description": "Retornar todos os cadastros de clientes",
                "operationId": "index_clientes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar novo cliente",
                "operationId": "store_clientes",
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cnpj",
                                    "data_fundacao"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "81.466.424/0001-48"
                                    },
                                    "data_fundacao": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2022-11-26"
                                    },
                                    "grupo_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/clientes/{id}": {
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Alterar cadastro do cliente",
                "operationId": "update_clientes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cnpj",
                                    "data_fundacao"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "81.466.424/0001-48"
                                    },
                                    "data_fundacao": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2022-11-26"
                                    },
                                    "grupo_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir cadastro de um cliente",
                "operationId": "destroy_clientes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}